FIX: 로직 및 버그 수정
1. 음악채널 초기화 후에 노래 명령어들이 동작합니다.
2. 로깅을 보충했습니다.
3. 사용법 명령어를 추가했습니다.

+++ src/cogs/music/music.py
@@ -87,7 +87,7 @@ class Music(commands.GroupCog, name="노래"):
 
     @app_commands.command(name="초기화", description="음악봇이 이상할 때, 새로 초기화합니다.")
     async def reset_music_channel_command(self, interaction: Interaction):
-        if not interaction.guild_id in self.bot.players:
+        if self.is_init_music_channel(interaction):
             return await interaction.response.send_message(
                 "`시작` 명령어로 음악 채널을 먼저 설정해주세요."
             )
@@ -96,6 +96,10 @@ class Music(commands.GroupCog, name="노래"):
     @app_commands.command(name="재생", description="노래를 재생합니다.")
     @app_commands.rename(title="제목or링크")
     async def add_song_to_player_command(self, interaction: Interaction, title: str):
+        if self.is_init_music_channel(interaction):
+            return await interaction.response.send_message(
+                "`시작` 명령어로 음악 채널을 먼저 설정해주세요."
+            )
         await interaction.response.defer(thinking=True)
         # await interaction.delete_original_response()
         await (await interaction.original_response()).delete()
@@ -107,14 +111,25 @@ class Music(commands.GroupCog, name="노래"):
 
     @app_commands.command(name="스킵", description="노래를 스킵합니다.")
     async def skip_song_command(self, interaction: Interaction):
+        if self.is_init_music_channel(interaction):
+            return await interaction.response.send_message(
+                "`시작` 명령어로 음악 채널을 먼저 설정해주세요."
+            )
         await self.bot.players[interaction.guild_id].skip(interaction)
 
     @app_commands.command(name="현재", description="현재 재생중인 노래 정보를 가져옵니다.")
     async def current_song_command(self, interaction: Interaction):
+        if self.is_init_music_channel(interaction):
+            return await interaction.response.send_message(
+                "`시작` 명령어로 음악 채널을 먼저 설정해주세요."
+            )
         song = self.bot.players[interaction.guild_id].current_song
         if song:
             await interaction.response.send_message(content=f"{song}", ephemeral=True)
             await self.bot.players[interaction.guild_id].youtube(interaction)
+
+    async def is_init_music_channel(self, interaction: Interaction):
+        return interaction.guild_id in self.bot.players
 
     async def initialize(self, guild: Guild, channel: TextChannel):
         if guild in self.bot.players:
